backend/api/serializers/EmploymentSerializer.py
from rest_framework import serializers
from api.models import Employment


class EmploymentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employment
        fields = "__all__"