backend/requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==4.2.4
django-countries==7.5.1
django-money==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.7
py-moneyed==3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
pyuca==1.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
